Add files via upload
A Charity Management system need to manage the information
about donors, donations and other relevant data. System need to
maintain information about each donors. System should able to
add new donor, remove the donor, update the information of the
donor, search the specific donor also it should display the
information of all the donor which are present in the database. Use
the concept of the Linked List to implement the above system.